requirements.txt
pandas==2.2.2
google-cloud-storage==2.16.0
google-cloud-secret-manager==2.17.0
pyarrow==16.0.0
sqlalchemy==2.0.29
pymssql==2.2.0
gunicorn==22.0.0